Create basic junction, pump and valve classes
